Territorios do saber -  Agrupar componentes curriculares
AB#11349 AB#11350 AB#11351 AB#11352 AB#11359 AB#11772 AB#11773 AB#11774

+++ src/SME.SGP.Aplicacao/Comandos/ComandosPlanoAula.cs
@@ -145,7 +145,7 @@ namespace SME.SGP.Aplicacao
                 // Caso a disciplina não possui vinculo com Jurema, os objetivos não devem ser exigidos
                 if (!permitePlanoSemObjetivos)
                 {
-                    permitePlanoSemObjetivos = !(await consultasObjetivoAprendizagem.DisciplinaPossuiObjetivosDeAprendizagem(Convert.ToInt32(aula.DisciplinaId)));
+                    permitePlanoSemObjetivos = !(await consultasObjetivoAprendizagem.DisciplinaPossuiObjetivosDeAprendizagem(Convert.ToInt64(aula.DisciplinaId)));
                 }
 
                 if (!permitePlanoSemObjetivos)
